Remove usages of internal Gradle API FileOperations
