+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/desktop/application/tasks/AbstractCheckNativeDistributionRuntime.kt
@@ -7,16 +7,16 @@ package org.jetbrains.compose.desktop.application.tasks
 
 import org.gradle.api.file.Directory
 import org.gradle.api.file.RegularFile
-import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.*
 import org.jetbrains.compose.desktop.application.internal.JvmRuntimeProperties
-import org.jetbrains.compose.desktop.application.internal.executableName
-import org.jetbrains.compose.desktop.application.internal.ioFile
-import org.jetbrains.compose.desktop.application.internal.notNullProperty
+import org.jetbrains.compose.internal.utils.executableName
+import org.jetbrains.compose.internal.utils.ioFile
+import org.jetbrains.compose.internal.utils.notNullProperty
 import org.jetbrains.compose.desktop.application.internal.ExternalToolRunner
 import org.jetbrains.compose.desktop.tasks.AbstractComposeDesktopTask
+import org.jetbrains.compose.internal.utils.clearDirs
 import java.io.File
 
 // __COMPOSE_NATIVE_DISTRIBUTIONS_MIN_JAVA_VERSION__
@@ -88,7 +88,7 @@ abstract class AbstractCheckNativeDistributionRuntime : AbstractComposeDesktopTa
     }
 
     private fun getJavaRuntimeVersionUnsafe(): String? {
-        cleanDirs(workingDir)
+        fileOperations.clearDirs(workingDir)
         val workingDir = workingDir.ioFile
 
         val printJavaRuntimeClassName = "PrintJavaRuntimeVersion"
