+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/desktop/application/tasks/AbstractNativeMacApplicationPackageTask.kt
@@ -10,9 +10,8 @@ import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.*
-import org.jetbrains.compose.desktop.application.internal.ioFile
-import org.jetbrains.compose.desktop.application.internal.notNullProperty
 import org.jetbrains.compose.desktop.tasks.AbstractComposeDesktopTask
+import org.jetbrains.compose.internal.utils.*
 import java.io.File
 
 abstract class AbstractNativeMacApplicationPackageTask : AbstractComposeDesktopTask() {
@@ -34,7 +33,7 @@ abstract class AbstractNativeMacApplicationPackageTask : AbstractComposeDesktopT
 
     @TaskAction
     fun run() {
-        cleanDirs(destinationDir, workingDir)
+        fileOperations.clearDirs(destinationDir, workingDir)
 
         createPackage(
             destinationDir = destinationDir.ioFile,
