+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/desktop/application/tasks/AbstractSuggestModulesTask.kt
@@ -14,12 +14,10 @@ import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.*
 import org.jetbrains.compose.desktop.application.dsl.DEFAULT_RUNTIME_MODULES
 import org.jetbrains.compose.desktop.application.internal.ComposeProperties
-import org.jetbrains.compose.desktop.application.internal.ioFile
 import org.jetbrains.compose.desktop.application.internal.ExternalToolRunner
-import org.jetbrains.compose.desktop.application.internal.jvmToolFile
 import org.jetbrains.compose.desktop.application.internal.files.normalizedPath
-import org.jetbrains.compose.desktop.application.internal.notNullProperty
 import org.jetbrains.compose.desktop.tasks.AbstractComposeDesktopTask
+import org.jetbrains.compose.internal.utils.*
 
 abstract class AbstractSuggestModulesTask : AbstractComposeDesktopTask() {
     @get:Input
@@ -47,7 +45,7 @@ abstract class AbstractSuggestModulesTask : AbstractComposeDesktopTask() {
     fun run() {
         val jtool = jvmToolFile("jdeps", javaHome = javaHome)
 
-        cleanDirs(workingDir)
+        fileOperations.clearDirs(workingDir)
         val args = arrayListOf<String>().apply {
             add("--print-module-deps")
             add("--ignore-missing-deps")
