+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/internal/utils/diagnosticUtils.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.compose.desktop.application.internal
+package org.jetbrains.compose.internal.utils
 
 import java.io.PrintWriter
 import java.io.StringWriter
