+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/internal/utils/osUtils.kt
@@ -1,9 +1,9 @@
 /*
- * Copyright 2020-2022 JetBrains s.r.o. and respective authors and developers.
+ * Copyright 2020-2023 JetBrains s.r.o. and respective authors and developers.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.compose.desktop.application.internal
+package org.jetbrains.compose.internal.utils
 
 import org.gradle.api.provider.Provider
 import org.jetbrains.compose.desktop.application.internal.files.checkExistingFile
