+++ gradle-plugins/compose/src/main/kotlin/org/jetbrains/compose/internal/utils/stringUtils.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package org.jetbrains.compose.internal
+package org.jetbrains.compose.internal.utils
 
 internal fun String.uppercaseFirstChar(): String =
     transformFirstCharIfNeeded(
